refactor(web): extract getServiceAccountLink to utils
